[pldmd] Remove yield_context storing as private
Storing yield_context is not safe. Generally it is passed on the
stack so that the currently executing coroutine is the only thing
to access it. By storing it as part of the object, is it possible
that this context would ever get used in another coroutine.

Thus remove the same from class.

Tested:
PLDM commands are successfully executed.

Change-Id: I974fb00cee57071f1d5c450a1ae0e323cc922823
Signed-off-by: Arun P. Mohanan <arun.p.m@linux.intel.com>