[pldmd] Rename sensor/effecter class names
The M&C code is structured in such a way that each sensor or effecter
class will handle only one instance of the same. In that case naming
the class with manager suffix is inappropriate.

Rename the same as Handler. Update function and variable names to
keep consistancy across sensors and effecters.

Tested:
Platform M&C init success.

Change-Id: Iaed55424a83e7479caef83e6d549bb8b8c22fcaa
Signed-off-by: Arun P. Mohanan <arun.p.m@linux.intel.com>