pldm: Minor fix in doCancelUpdateComponent() call
CancelUpdateComponent command will be called if TransferComplete or
VerifyComplete command is failed in the update process. Since same
return value is used for doCancelUpdateComponent() and if
CancelUpdateComponent is success, it gives wrong impression to the upper
layers that FW update is success.

Now using different return value for doCancelUpdateComponent().

Test:
Expected return value is returned during FW update process.

Change-Id: I19bb677d69776d8a9f7f5f6534535380227c3b18
Signed-off-by: Gade-RajasekharReddy@ <raja.sekhar.reddy.gade@intel.com>