[PLDMD] Call pldm_pdr_add with uint32 as size
The spec of PLDM states that pdr repositories
are to be described as uint32. However, in standard
C++ the size of vectors is described as size_t.

Given that size_t may be represented as either 32 or 64 bit,
the size of a vector of pdr repositories needs to be
downcasted to uint32_t to conform to the spec.

Tested:
 - Compiles and works for arm32 and x86_64

Change-Id: I3327745783ec9f79a8a8a6876cc3621dd15aa3cd
Signed-off-by: Sandino Flores <sandino.flores.moreno@intel.com>