[pldmd] Discard PDR record exceeding size limit
When large PDR record exceeding the size limit is received, the
multipart transfer loop is exited but the record will still be added
to the repository if transferComplete is set. transferComplete is
always set for singlepart transfers or as long as the CRC matches
for a multipart in handleGetPDRResp.

Clear transferComplete flag to mitigate the above issue.

Tested:
1) Verified getPDR worked as expected
2) PLDM M&C init success

Change-Id: I4d88557fc0b1662bdc332d751ced80d19836f0c1
Signed-off-by: Arun P. Mohanan <arun.p.m@linux.intel.com>