pldmd: Updated the termination handler
Signal handlers are implemented in pldmd for handling
termination and interrupt signal. This handler is
executed when pldmd is manually stopped.

The cleanest way for a handler to terminate
the process is to raise the same signal that
ran the handler in the first place. while
raising the signal at the end with SIG_DFL
kernel will execute default action associated
with the signal

https://www.gnu.org/software/libc/manual/html_node/Termination-in-Handler.html

Tested:
pldmd is manually stopped and no undefined
behavior were observed
stoped mctpd first and then stoped pldmd and
no undefined behavior were observed

Change-Id: If5d352d7e1191d31e0d31531c4f3849c6f23a617
Signed-off-by: gokulsanker <gokul.sanker.v.g@intel.com>