pldmd: Updated code to set component info to D-bus
addCompDataToDBus() expose the component set info
into the D-bus. Datatypes of few parameters were
changed to bitfield and this change was not added to
addCompDataToDBus().It cause an exception while
updating the D-bus. This is resolved by updating
the datatypes properly

Tested:
Tested in device and D-bus were exposed properly

Change-Id: Idb866569c0c347f1069bccf77ab5460aeb063181
Signed-off-by: gokulsanker <gokul.sanker.v.g@intel.com>