[PLDM]State Sensor Interface not exposed on D-Bus
State Sensor interface is not exposed on D-Bus
and State Sensor properties are not registered with D-Bus .

During Sensor polling, the sensorErrorCheck() is called and it is failed
due to the uninilialized value of errCount(due to dummy value in
errorCount was more than thresholds). Because of which
sensorErrorCheck() is continued to be called repeatatively and hence
causing further functions to populate sensor values and registering
sensor interfaces and properties to D-Bus doesn't happen.

Initialization of errCount is taken care in other files also (State
Effecters, Numeric Effecters and Numeric Sensors).

Log level changed to expected value in incrementError() API.

TESTED: Verified that the State Sensor values are populated.
State Sensor Interface and the properties are registered to D-Bus.

Change-Id: I25b68068add66a27f899666e576d072d6adcf338
Signed-off-by: Ashwini N S <ashwinix.niduvani.shivashankarappa@intel.com>