pldmd: Fix in RequestUpdate command
Total number of components is passed in RequestUpdate commands request
instead of applicable components count.
Eg: PLDM package can have 4 components in it and if 2 are applicable
components UA(BMC) need to pass applicable components count(2) in the
request update command.
This patch fixes it.

Tested:
PLDM firmware update is success for add-in card.

Change-Id: I99671486be4d4ddf842bb9bb0569f04711a528a8
Signed-off-by: Gade-RajasekharReddy@ <raja.sekhar.reddy.gade@intel.com>