pldm: Fix for PLDM_FWU progress in redfish
At present the progress percentage of PLDM firmware update in redfish is
updated only for updatable components. There could be some components
which can not be updated for valid reasons. The progress percentage
need to be updated for such components as well.

Tested:
Proper progress % is updated in redfish during PLDM FWU.

Change-Id: Ie5d686eca2ae9fef99238120a840cb328cc53233
Signed-off-by: Gade-RajasekharReddy@ <raja.sekhar.reddy.gade@intel.com>