[pldmd] Expose PLDM sensors as baseboard sensors
This patch creates association to the baseboard object for numeric PLDM
sensors. With this association created, numeric sensors can be viewed
in Redfish as baseboard chassis sensors.

Resolves "TODO: Support to update associations".
Also adds the missing "Unit" DBus property of numeric sensor.

This feature is controlled by the CMake option EXPOSE_BASEBOARD_SENSOR,
which is disabled by default.

Tested:
When compiled with EXPOSE_BASEBOARD_SENSOR=ON, the PLDM sensor info
can be listed and read in /redfish/v1/Chassis/<baseboard>/Sensors.

Change-Id: I4d9801a4bc1a7cc2d1399588197674159789efc1
Signed-off-by: Jiaqing Zhao <jiaqing.zhao@intel.com>