[pldmd] Expose PLDM device as chassis in Redfish
This patch creates an inventory object for each PLDM device, and
associate its sensors to the object. In this case, it is exposed as a
standalone chassis in Redfish.

This feature is controlled by the CMake option EXPOSE_CHASSIS, which is
disabled by default, and this option cannot be enabled together with
EXPOSE_BASEBOARD_SENSOR.

Tested:
When compiled with EXPOSE_CHASSIS=ON, the PLDM device is listed in
/redfish/v1/Chassis/<DeviceName_TID>, and its sensors can be listed
and read in /redfish/v1/Chassis/<DeviceName_TID>/Sensors.

Change-Id: I76bfed188914fa91d9dc4ca176948de1c77f6697
Signed-off-by: Jiaqing Zhao <jiaqing.zhao@intel.com>