pldmd: Fix in CancelUpdate cmd call flow (#14) (#13)
This is porting of main line PR #13 

If PLDM firmware update failed for all the components,
CancelUpdateComponent should be called for all failed components and then
CancelUpdate command should be called. Since same return value is
used for calling both the commands, CancelUpdate is not called when
CancelUpdateComponent is success. This patch fixes it.

Tested:
Unit tested the code and it is working as expected.

Signed-off-by: Gade-RajasekharReddy <raja.sekhar.reddy.gade@intel.com>